+++ code-api/Dockerfile
@@ -1,0 +1,13 @@
+FROM node:18-alpine
+
+WORKDIR /usr/src/app
+
+COPY package*.json .
+
+RUN npm install
+
+COPY . .
+
+EXPOSE 9000
+
+CMD ["npm", "run", "dev"]
