+++ code-api/Makefile
@@ -1,0 +1,2 @@
+build:
+	docker build -t code-api:peerprep .
