+++ code-api/nodemon.json
@@ -1,0 +1,6 @@
+{
+	"ignore": ["**/*.test.ts", "**/*.spec.ts", "node_modules"],
+	"watch": ["src"],
+	"exec": "npm start",
+	"ext": "ts"
+}
